requirements.txt
adal==1.2.7
aiohttp==3.7.4.post0
aiounittest==1.3.0
async-timeout==3.0.1
attrs==21.4.0
azure-cognitiveservices-language-luis==0.2.0
azure-common==1.1.28
babel==2.7.0
botbuilder-ai==4.14.2
botbuilder-core==4.14.2
botbuilder-dialogs==4.14.2
botbuilder-integration-aiohttp==4.14.2
botbuilder-schema==4.14.2
botframework-connector==4.14.2
botframework-streaming==4.14.2
certifi==2022.5.18.1
cffi==1.15.0
chardet==4.0.0
cryptography==3.4.8
datedelta==1.4
emoji==1.7.0
grapheme==0.6.0
idna==2.10
isodate==0.6.1
jsonpickle==1.4.2
msal==1.6.0
msrest==0.6.19
msrestazure==0.6.4
multidict==6.0.2
multipledispatch==0.6.0
oauthlib==3.2.0
pycparser==2.21
pyjwt==1.7.1
pyodbc==4.0.32
python-dateutil==2.8.2
pytz==2022.1
recognizers-text==1.0.2a2
recognizers-text-choice==1.0.2a2
recognizers-text-date-time==1.0.2a2
recognizers-text-number==1.0.2a2
recognizers-text-number-with-unit==1.0.2a2
regex==2019.8.19
requests==2.25.1
requests-oauthlib==1.3.1
six==1.16.0
typing-extensions==4.2.0
urllib3==1.26.9
yarl==1.4.2